backend/requirements.txt
# requirements.txt
Flask
Flask-Cors
PyJWT
python-dotenv